move all dashboard-related components to Dashboard.js
